adding an annotation vs assembly id

+++ structural-annotation.wdl
@@ -21,20 +21,21 @@ workflow s_annotate {
   String  database_location
   String  container
   String? gm_license
+  
 
-  if(pre_qc_execute) {
-    call pre_qc {
-      input:
-        project_type = imgap_project_type,
-        input_fasta = imgap_input_fasta,
-        project_id = imgap_project_id,
-        container=container
-    }
+  call pre_qc {
+    input:
+      pre_qc_execute = pre_qc_execute,
+      project_type = imgap_project_type,
+      input_fasta = imgap_input_fasta,
+      project_id = imgap_project_id,
+      container=container
   }
+
   if(trnascan_se_execute) {
     call trnascan.trnascan {
       input:
-        imgap_input_fasta = imgap_input_fasta,
+        imgap_input_fasta = pre_qc.fasta,
         imgap_project_id = imgap_project_id,
         imgap_project_type = imgap_project_type,
         additional_threads = additional_threads,
@@ -45,7 +46,7 @@ workflow s_annotate {
     call rfam.rfam {
       input:
         cmzscore = cmzscore,
-        imgap_input_fasta = imgap_input_fasta,
+        imgap_input_fasta = pre_qc.fasta,
         imgap_project_id = imgap_project_id,
         imgap_project_type = imgap_project_type,
         database_location = database_location,
@@ -56,7 +57,7 @@ workflow s_annotate {
   if(crt_execute) {
     call crt.crt {
       input:
-        imgap_input_fasta = imgap_input_fasta,
+        imgap_input_fasta = pre_qc.fasta,
         imgap_project_id = imgap_project_id,
         container=container
     }
@@ -65,7 +66,7 @@ workflow s_annotate {
    if(cds_prediction_execute) {
      call cds_prediction.cds_prediction {
        input:
-         imgap_input_fasta = imgap_input_fasta,
+         imgap_input_fasta = pre_qc.fasta,
          imgap_project_id = imgap_project_id,
          imgap_project_type = imgap_project_type,
          prodigal_execute = prodigal_execute,
@@ -79,7 +80,7 @@ workflow s_annotate {
 
     call gff_merge {
       input:
-        input_fasta = imgap_input_fasta,
+        input_fasta = pre_qc.fasta,
         project_id = imgap_project_id,
         rfam_gff = rfam.rfam_gff,
         trna_gff = trnascan.gff,
@@ -96,7 +97,7 @@ workflow s_annotate {
   if(prodigal_execute || genemark_execute)  {
     call fasta_merge {
       input:
-       # input_fasta = imgap_input_fasta,
+       # input_fasta = pre_qc.fasta,
         project_id = imgap_project_id,
         final_gff = gff_merge.final_gff,
         cds_genes = cds_prediction.genes,
@@ -108,7 +109,7 @@ workflow s_annotate {
   if(gff_and_fasta_stats_execute) {
     call gff_and_fasta_stats {
       input:
-        input_fasta = imgap_input_fasta,
+        input_fasta = pre_qc.fasta,
         project_id = imgap_project_id,
         final_gff = gff_merge.final_gff,
         container = container
@@ -117,7 +118,7 @@ workflow s_annotate {
   if(imgap_project_type == "isolate") {
     call post_qc {
       input:
-        input_fasta = imgap_input_fasta,
+        input_fasta = pre_qc.fasta,
         project_id = imgap_project_id,
         container = container
     }
@@ -144,15 +145,17 @@ workflow s_annotate {
     String? rfam_version = rfam.rfam_version
     File? proteins = fasta_merge.final_proteins
     File? genes = fasta_merge.final_genes
+    File? map_file = pre_qc.map_file
   }
 }
 
 task pre_qc {
+  Boolean pre_qc_execute
   String bin="/opt/omics/bin/qc/pre-annotation/fasta_sanity.py"
   String project_type
   File   input_fasta
   String project_id
-  String rename = "yes"
+  Boolean rename = true
   Float  n_ratio_cutoff = 0.5
   Int    seqs_per_million_bp_cutoff = 500
   Int    min_seq_length = 150
@@ -162,43 +165,49 @@ task pre_qc {
     set -euo pipefail
     tmp_fasta="${input_fasta}.tmp"
     qced_fasta="${project_id}_contigs.fna"
-    grep -v '^\s*$' ${input_fasta} | tr -d '\r' | \
-    sed 's/^>[[:blank:]]*/>/g' > $tmp_fasta
-    acgt_count=`grep -v '^>' $tmp_fasta | grep -o [acgtACGT] | wc -l`
-    n_count=`grep -v '^>' $tmp_fasta | grep -o '[^acgtACGT]' | wc -l`
-    n_ratio=`echo $n_count $acgt_count | awk '{printf "%f", $1 / $2}'`
-    if (( $(echo "$n_ratio >= ${n_ratio_cutoff}" | bc) ))
+    if [[ "${pre_qc_execute}" = true ]]
     then
-        rm $tmp_fasta
-        exit 1
-    fi
+      grep -v '^\s*$' ${input_fasta} | tr -d '\r' | \
+      sed 's/^>[[:blank:]]*/>/g' > $tmp_fasta
+      acgt_count=`grep -v '^>' $tmp_fasta | grep -o [acgtACGT] | wc -l`
+      n_count=`grep -v '^>' $tmp_fasta | grep -o '[^acgtACGT]' | wc -l`
+      n_ratio=`echo $n_count $acgt_count | awk '{printf "%f", $1 / $2}'`
+      if (( $(echo "$n_ratio >= ${n_ratio_cutoff}" | bc) ))
+      then
+          rm $tmp_fasta
+          exit 1
+      fi
 
-    if [[ ${project_type} == "isolate" ]]
-    then
-        seq_count=`grep -c '^>' $tmp_fasta`
-        bp_count=`grep -v '^>' $tmp_fasta | tr -d '\n' | wc -m`
-        seqs_per_million_bp=$seq_count
-        if (( $bp_count > 1000000 ))
-        then
-            divisor=$(echo $bp_count | awk '{printf "%.f", $1 / 1000000}')
-            seqs_per_million_bp=$(echo $seq_count $divisor | \
-                                  awk '{printf "%.2f", $1 / $2}')
-        fi
-        if (( $(echo "$seqs_per_million_bp > ${seqs_per_million_bp_cutoff}" | bc) ))
-        then
-            rm $tmp_fasta
-            exit 1
-        fi
-    fi
+      if [[ ${project_type} == "isolate" ]]
+      then
+          seq_count=`grep -c '^>' $tmp_fasta`
+          bp_count=`grep -v '^>' $tmp_fasta | tr -d '\n' | wc -m`
+          seqs_per_million_bp=$seq_count
+          if (( $bp_count > 1000000 ))
+          then
+              divisor=$(echo $bp_count | awk '{printf "%.f", $1 / 1000000}')
+              seqs_per_million_bp=$(echo $seq_count $divisor | \
+                                    awk '{printf "%.2f", $1 / $2}')
+          fi
+          if (( $(echo "$seqs_per_million_bp > ${seqs_per_million_bp_cutoff}" | bc) ))
+          then
+              rm $tmp_fasta
+              exit 1
+          fi
+      fi
 
-    fasta_sanity_cmd="${bin} $tmp_fasta $qced_fasta"
-    if [[ ${rename} == "yes" ]]
-    then
-        fasta_sanity_cmd="$fasta_sanity_cmd -p ${project_id}"
+      fasta_sanity_cmd="${bin} $tmp_fasta $qced_fasta"
+      if [[ "${rename}" = true ]]
+      then
+          fasta_sanity_cmd="$fasta_sanity_cmd -p ${project_id}"
+      fi
+      fasta_sanity_cmd="$fasta_sanity_cmd -l ${min_seq_length}"
+      ${bin} -v 
+      $fasta_sanity_cmd
+      rm $tmp_fasta
+    else
+      cp ${input_fasta} $qced_fasta
     fi
-    fasta_sanity_cmd="$fasta_sanity_cmd -l ${min_seq_length}"
-    $fasta_sanity_cmd
-    rm $tmp_fasta
   >>>
 
   runtime {
@@ -209,6 +218,8 @@ task pre_qc {
     
   output {
     File fasta = "${project_id}_contigs.fna"
+    File? map_file = "${project_id}_contig_names_mapping.tsv"
+    File  out_log = stdout()
   }
 }
